Dependency resolution fixes (#301)
## What?
* Adds an incoming transition to _binary/_library rules that transitions to the highest compatible framework in the `target_frameworks`
* Adds a `private_deps` attribute to _binary/_library rules that is intended for compile time dependencies. These dependencies are not transitively passed and do not become part of the runfiles.
* Some cleanup and reorganizing

## Why?
The transition is needed because before the change we were just taking the default tfm and passing it to the build actions. This did not make sense since you could update rules_dotnet and then the default tfm would be 6.0 but your targets would still target 5.0. There was also an issue that if you were e.g. targeting `netstandard2.1` we would still pass the default tfm to the build actions. 

The `private_deps` attribute is needed so that we can pass in targeting packs without passing them transitively to parents. Targeting packs should not be passed transitively because that causes conflicts between targeting packs. 

Fixes #297