Fix runtimeconfig for AspNetCore projects (#323)
* Fix runtimeconfig for AspNetCore projects

## What?
Adds a a new `project_sdk` attribute to all targets that defaults to `default`
The attribute can also be set to `web` which indicates that the Web sdk is being targeted
and thus the runtimeconfig will include the AspNetCore framework in the `includedFrameworks/frameworks` key of the config.

## Why?
This is required to make the binary targets runnable. At some point I hope to simplify this so that
the correct NuGet packages are implicitly added to the targets without having to declare them in
`private_deps` or `runtime_packs`.